+++ harry_potter/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "tsc -b && vite build",
+    "build": "tsc -b && vite build && cp ./src/_redirects ./dist/_redirects",
     "format:fix": "prettier --write .",
     "lint": "eslint .",
     "preview": "vite preview",
