+++ harry_potter/src/api/index.ts
@@ -1,1 +1,1 @@
-export * from './apiRequest.tsx';
+export { potterApi } from './api_request.tsx';
