+++ harry_potter/src/components/ErrorBoundary/index.ts
@@ -1,1 +1,1 @@
-export { ErrorBoundary } from './errorBoundary.tsx';
+export { ErrorBoundary } from './error_boundary.tsx';
