+++ harry_potter/src/components/ErrorModal/error_window.tsx
