+++ harry_potter/src/components/Results/card_list.tsx
