+++ harry_potter/src/components/index.ts
@@ -1,12 +1,12 @@
-export * from './Card';
-export * from './ErrorBoundary';
-export * from './ErrorModal';
-export * from './Footer';
-export * from './Header';
-export * from './Pangination';
-export * from './Results';
-export * from './Search';
-export * from './Spinner';
-export * from './FlyoutElement';
-export * from './Settings';
-export * from './Router';
+export { Card } from './Card';
+export { ErrorBoundary } from './ErrorBoundary';
+export { ErrorModal } from './ErrorModal';
+export { Footer } from './Footer';
+export { Header } from './Header';
+export { Pangination } from './Pangination';
+export { CardList } from './Results';
+export { SearchFieldComponent } from './Search';
+export { Spinner } from './Spinner';
+export { FlyoutElement } from './FlyoutElement';
+export { Settings } from './Settings';
+export { Router } from './Router';
