+++ harry_potter/src/context/index.ts
@@ -1,3 +1,4 @@
-export * from './themeProvider.tsx';
-export * from './musicProvider.tsx';
-export * from './contexts.tsx';
+export { ThemeProvider } from './themeProvider.tsx';
+export { ThemeContext } from './contexts.tsx';
+export { SoundContext } from './contexts.tsx';
+export { SoundProvider } from './musicProvider.tsx';
