+++ harry_potter/src/data/index.ts
@@ -1,4 +1,4 @@
 export * from './store.tsx';
-export * from './storeSlice.tsx';
-import PotterReducer from './storeSlice.tsx';
+export * from './store_slice.tsx';
+import PotterReducer from './store_slice.tsx';
 export { PotterReducer };
