+++ harry_potter/src/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useSearchStringLS.tsx';
+export * from './use_storage.tsx';
