+++ harry_potter/src/pages/detailed/detailed_card.test.tsx
