+++ harry_potter/src/pages/mainPage/index.ts
@@ -1,1 +1,1 @@
-export * from './mainPage.tsx';
+export { MainPage } from './main_page.tsx';
