+++ harry_potter/src/pages/notFound/index.ts
@@ -1,1 +1,1 @@
-export * from './notFoundPage.tsx';
+export { NotFoundPage } from './not_found_page.tsx';
