+++ harry_potter/src/types/index.ts
@@ -1,1 +1,11 @@
-export * from './types.tsx';
+export type {
+  State,
+  StateProps,
+  ErrorDetails,
+  ErrorProps,
+  ErrorBoundaryState,
+  CharactersResponse,
+  CharacterResponse,
+  Character,
+  SearchPropsForCharacters,
+} from './types.tsx';
